requirements.txt
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
django-structlog==2.1.0
drf_yasg==1.20.0
markdown==3.3.3
django-filter==2.4.0
structlog==21.1.0
